requirements.txt
requests==2.31.0
boto3==1.34.0
social-post-api
python-dotenv==1.0.0
